:hammer: disable and enable tests based on env flags
