+++ src/test/debuggerTest.ts
@@ -9,7 +9,8 @@ process.env.CODE_TESTS_WORKSPACE = path.join(__dirname, '..', '..', 'src', 'test
 process.env.IS_CI_SERVER_TEST_DEBUGGER = '1';
 
 function start() {
-    console.log('start Multiroot Debugger tests');
+    console.log('*'.repeat(100));
+    console.log('Start Debugger tests');
     require('../../node_modules/vscode/bin/test');
 }
 start();
