+++ src/test/index.ts
@@ -3,22 +3,30 @@ if ((Reflect as any).metadata === undefined) {
     // tslint:disable-next-line:no-require-imports no-var-requires
     require('reflect-metadata');
 }
-import { workspace } from 'vscode';
 import { MochaSetupOptions } from 'vscode/lib/testrunner';
+import { IS_CI_SERVER, IS_CI_SERVER_TEST_DEBUGGER, IS_MULTI_ROOT_TEST } from './constants';
 import * as testRunner from './testRunner';
 
 process.env.VSC_PYTHON_CI_TEST = '1';
-process.env.IS_MULTI_ROOT_TEST = (Array.isArray(workspace.workspaceFolders) && workspace.workspaceFolders.length > 1);
+process.env.IS_MULTI_ROOT_TEST = IS_MULTI_ROOT_TEST;
+
+let grep: string | undefined;
+// If running on CI server and we're running the debugger tests, then ensure we only run debug tests.
+// We do this to ensure we only run debugger test, as debugger tests are very flaky on CI.
+// So the solution is to run them separately and first on CI.
+if (IS_CI_SERVER && IS_CI_SERVER_TEST_DEBUGGER) {
+    grep = 'Debug';
+}
 
 // You can directly control Mocha options by uncommenting the following lines.
 // See https://github.com/mochajs/mocha/wiki/Using-mocha-programmatically#set-options for more info.
 // Hack, as retries is not supported as setting in tsd.
-// tslint:disable-next-line:no-any
 const options: MochaSetupOptions & { retries: number } = {
     ui: 'tdd',
     useColors: true,
     timeout: 25000,
-    retries: 3
+    retries: 3,
+    grep
 };
 testRunner.configure(options, { coverageConfig: '../coverconfig.json' });
 module.exports = testRunner;
